+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Note: accepted uncompressed archive total size is increased to 256 MiB.
 
 ## `20251120t084200-all`
  * Bump runtimeVersion to `2025.11.18`.